Revert "Update sherlock-platform to IntelliJ idea/251.26094.121 (#166)"
This reverts commit 854ce640269d6be52028226f600fe66fea5b2135.

+++ java/java-impl-refactorings/src/com/intellij/refactoring/convertToInstanceMethod/ConvertToInstanceMethodProcessor.java
@@ -1,4 +1,4 @@
-// Copyright 2000-2025 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.intellij.refactoring.convertToInstanceMethod;
 
 import com.intellij.codeInsight.ChangeContextUtil;
@@ -27,7 +27,6 @@ import com.intellij.usageView.UsageViewDescriptor;
 import com.intellij.util.IncorrectOperationException;
 import com.intellij.util.VisibilityUtil;
 import com.intellij.util.containers.MultiMap;
-import com.siyeh.ig.psiutils.VariableAccessUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -39,7 +38,6 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
   private static final Logger LOG = Logger.getInstance(ConvertToInstanceMethodProcessor.class);
   private PsiMethod myMethod;
   private @Nullable PsiParameter myTargetParameter;
-  private final boolean myParameterIsReassigned;
   private PsiClass myTargetClass;
   private Map<PsiTypeParameter, PsiTypeParameter> myTypeParameterReplacements;
   private static final Key<PsiTypeParameter> BIND_TO_TYPE_PARAMETER = Key.create("REPLACEMENT");
@@ -49,13 +47,11 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
 
   public ConvertToInstanceMethodProcessor(final Project project,
                                           final PsiMethod method,
-                                          final @Nullable PsiParameter targetParameter,
+                                          @Nullable final PsiParameter targetParameter,
                                           final String newVisibility) {
     super(project);
     myMethod = method;
     myTargetParameter = targetParameter;
-    myParameterIsReassigned = targetParameter != null && 
-                              VariableAccessUtils.variableIsAssigned(myTargetParameter, myTargetParameter.getDeclarationScope());
     LOG.assertTrue(method.hasModifierProperty(PsiModifier.STATIC));
     if (myTargetParameter != null) {
       LOG.assertTrue(myTargetParameter.getDeclarationScope() == myMethod);
@@ -76,7 +72,8 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
   }
 
   @Override
-  protected @NotNull UsageViewDescriptor createUsageViewDescriptor(UsageInfo @NotNull [] usages) {
+  @NotNull
+  protected UsageViewDescriptor createUsageViewDescriptor(UsageInfo @NotNull [] usages) {
     return new MoveInstanceMethodViewDescriptor(myMethod, myTargetParameter, myTargetClass);
   }
 
@@ -113,7 +110,7 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     }
 
     if (myTargetParameter != null) {
-      for (final PsiReference ref : ReferencesSearch.search(myTargetParameter, new LocalSearchScope(myMethod), false).asIterable()) {
+      for (final PsiReference ref : ReferencesSearch.search(myTargetParameter, new LocalSearchScope(myMethod), false)) {
         final PsiElement element = ref.getElement();
         if (element instanceof PsiReferenceExpression || element instanceof PsiDocParamRef) {
           result.add(new ParameterUsageInfo(ref));
@@ -206,7 +203,8 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     EditorHelper.openInEditor(doRefactoring(usages));
   }
 
-  private @NotNull PsiMethod doRefactoring(UsageInfo[] usages) {
+  @NotNull
+  private PsiMethod doRefactoring(UsageInfo[] usages) {
     myTypeParameterReplacements = buildTypeParameterReplacements();
     List<PsiClass> inheritors = new ArrayList<>();
 
@@ -217,7 +215,7 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
       if (usage instanceof MethodCallUsageInfo) {
         processMethodCall(((MethodCallUsageInfo)usage).getMethodCall());
       }
-      else if (usage instanceof ParameterUsageInfo && !myParameterIsReassigned) {
+      else if (usage instanceof ParameterUsageInfo) {
         processParameterUsage((ParameterUsageInfo)usage);
       }
       else if (usage instanceof ImplementingClassUsageInfo) {
@@ -229,20 +227,7 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     }
 
     prepareTypeParameterReplacement();
-    if (myTargetParameter != null) {
-      if (myParameterIsReassigned) {
-        PsiDeclarationStatement statement =
-          JavaPsiFacade.getElementFactory(myProject).createVariableDeclarationStatement(myTargetParameter.getName(),
-                                                                                        myTargetParameter.getType(),
-                                                                                        createThisExpression());
-        PsiCodeBlock body = myMethod.getBody();
-        assert body != null;
-        PsiElement first = body.getFirstBodyElement();
-        assert first != null;
-        first.getParent().addBefore(statement, first);
-      }
-      myTargetParameter.delete();
-    }
+    if (myTargetParameter != null) myTargetParameter.delete();
     ChangeContextUtil.encodeContextInfo(myMethod, true);
     PsiMethod result;
     if (!myTargetClass.isInterface()) {
@@ -261,13 +246,12 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
       if (!markAsDefault) {
         for (final PsiClass psiClass : inheritors) {
           final PsiMethod newMethod = addMethodToClass(psiClass);
-          String modifier = myNewVisibility != null && !myNewVisibility.equals(VisibilityUtil.ESCALATE_VISIBILITY)
-                            ? myNewVisibility
-                            : PsiModifier.PUBLIC;
-          PsiUtil.setModifierProperty(newMethod, modifier, true);
+          PsiUtil.setModifierProperty(newMethod, myNewVisibility != null && !myNewVisibility.equals(VisibilityUtil.ESCALATE_VISIBILITY) ? myNewVisibility
+                                                                                                                                        : PsiModifier.PUBLIC, true);
         }
       }
     }
+    myMethod.delete();
     return result;
   }
 
@@ -329,7 +313,7 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     if (myTypeParameterReplacements == null) return;
     final Collection<PsiTypeParameter> typeParameters = myTypeParameterReplacements.keySet();
     for (final PsiTypeParameter parameter : typeParameters) {
-      for (final PsiReference reference : ReferencesSearch.search(parameter, new LocalSearchScope(myMethod), false).asIterable()) {
+      for (final PsiReference reference : ReferencesSearch.search(parameter, new LocalSearchScope(myMethod), false)) {
         if (reference.getElement() instanceof PsiJavaCodeReferenceElement) {
           reference.getElement().putCopyableUserData(BIND_TO_TYPE_PARAMETER, myTypeParameterReplacements.get(parameter));
         }
@@ -342,16 +326,7 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
   }
 
   private PsiMethod addMethodToClass(final PsiClass targetClass) {
-    final PsiMethod newMethod;
-    if (targetClass == myMethod.getContainingClass()) {
-      newMethod = myMethod;
-    }
-    else {
-      newMethod = (PsiMethod)targetClass.add(myMethod);
-      PsiMethod copy = (PsiMethod)myMethod.copy();
-      myMethod.delete();
-      myMethod = copy;
-    }
+    final PsiMethod newMethod = (PsiMethod)targetClass.add(myMethod);
     final PsiModifierList modifierList = newMethod.getModifierList();
     modifierList.setModifierProperty(PsiModifier.STATIC, false);
     ChangeContextUtil.decodeContextInfo(newMethod, null, null);
@@ -495,11 +470,13 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
   }
 
   @Override
-  protected @NotNull String getCommandName() {
+  @NotNull
+  protected String getCommandName() {
     return ConvertToInstanceMethodHandler.getRefactoringName();
   }
 
-  public @Nullable Map<PsiTypeParameter, PsiTypeParameter> buildTypeParameterReplacements() {
+  @Nullable
+  public Map<PsiTypeParameter, PsiTypeParameter> buildTypeParameterReplacements() {
     if (myTargetParameter == null) {
       return Collections.emptyMap();
     }
@@ -508,9 +485,10 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     return calculateReplacementMap(substitutor, myTargetClass, myMethod);
   }
 
-  private static @Nullable Map<PsiTypeParameter, PsiTypeParameter> calculateReplacementMap(final PsiSubstitutor substitutor,
-                                                                                           final PsiClass targetClass,
-                                                                                           final PsiElement containingElement) {
+  @Nullable
+  private static Map<PsiTypeParameter, PsiTypeParameter> calculateReplacementMap(final PsiSubstitutor substitutor,
+                                                                                 final PsiClass targetClass,
+                                                                                 final PsiElement containingElement) {
     final HashMap<PsiTypeParameter, PsiTypeParameter> result = new HashMap<>();
     for (PsiTypeParameter classTypeParameter : PsiUtil.typeParametersIterable(targetClass)) {
       if (!(substitutor.substitute(classTypeParameter) instanceof PsiClassType classType)) return null;
@@ -526,7 +504,8 @@ public final class ConvertToInstanceMethodProcessor extends BaseRefactoringProce
     return myMethod;
   }
 
-  public @Nullable PsiParameter getTargetParameter() {
+  @Nullable
+  public PsiParameter getTargetParameter() {
     return myTargetParameter;
   }
 }
