+++ platform/platform-impl/src/com/intellij/ide/actions/RecentLocationsDataModel.kt
@@ -1,4 +1,4 @@
-// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+// Copyright 2000-2022 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.intellij.ide.actions
 
 import com.intellij.codeInsight.breadcrumbs.FileBreadcrumbsCollector
@@ -15,16 +15,20 @@ import com.intellij.openapi.util.text.StringUtil
 import com.intellij.util.DocumentUtil
 import com.intellij.util.concurrency.SynchronizedClearableLazy
 import com.intellij.util.containers.ContainerUtil
+import org.jetbrains.annotations.ApiStatus
 import org.jetbrains.annotations.Nls
+import java.util.*
+import java.util.function.Consumer
 import java.util.function.Function
+import java.util.stream.Collectors
 import kotlin.math.max
 import kotlin.math.min
 
-internal class RecentLocationsDataModel(
-  private val project: Project,
-  private val placesSupplier: Function<Boolean, List<IdeDocumentHistoryImpl.PlaceInfo>>?,
-  private val placesRemover: ((List<IdeDocumentHistoryImpl.PlaceInfo>) -> Unit)?,
-) {
+@ApiStatus.Internal
+internal class RecentLocationsDataModel(val project: Project,
+                                        private val placesSupplier: Function<in Boolean, out List<IdeDocumentHistoryImpl.PlaceInfo>>?,
+                                        private val placesRemover: Consumer<in List<IdeDocumentHistoryImpl.PlaceInfo>>?) {
+
   private val navigationPlaces: SynchronizedClearableLazy<List<RecentLocationItem>> = calculateItems(project, false)
 
   private val changedPlaces: SynchronizedClearableLazy<List<RecentLocationItem>> = calculateItems(project, true)
@@ -46,10 +50,9 @@ internal class RecentLocationsDataModel(
   }
 
   private fun collectBreadcrumbs(project: Project, items: List<RecentLocationItem>): Map<IdeDocumentHistoryImpl.PlaceInfo, @Nls String> {
-    return items
-      .asSequence()
+    return items.stream()
       .map(RecentLocationItem::info)
-      .associateBy(keySelector = { it }, valueTransform = { getBreadcrumbs(project, it) })
+      .collect(Collectors.toMap({ it }, { getBreadcrumbs(project, it) }))
   }
 
   @Nls
@@ -92,12 +95,10 @@ internal class RecentLocationsDataModel(
       val places = ContainerUtil.reverse(if (changed) IdeDocumentHistory.getInstance(project).changePlaces
                                          else IdeDocumentHistory.getInstance(project).backPlaces)
       for (place in places) {
-        if (result.none { IdeDocumentHistory.getInstance(project).isSame(place, it.info) }) {
+        if (result.stream().noneMatch { IdeDocumentHistory.getInstance(project).isSame(place, it.info) }) {
           result.add(newLocationItem(place) ?: continue)
         }
-        if (result.size >= maxPlaces) {
-          break
-        }
+        if (result.size >= maxPlaces) break
       }
     }
     return result
@@ -108,7 +109,6 @@ internal class RecentLocationsDataModel(
     if (positionOffset == null || !positionOffset.isValid) {
       return null
     }
-
     assert(positionOffset.startOffset == positionOffset.endOffset)
 
     val fileDocument = positionOffset.document
@@ -122,32 +122,18 @@ internal class RecentLocationsDataModel(
   }
 
   fun removeItems(project: Project, isChanged: Boolean, items: List<RecentLocationItem>) {
-    if (isChanged) {
-      changedPlaces.drop()
-    }
-    else {
-      navigationPlaces.drop()
-    }
-
-    placesRemover?.let {
-      it(items.map { it.info })
+    if (isChanged) changedPlaces.drop() else navigationPlaces.drop()
+    if (placesRemover != null) {
+      placesRemover.accept(ContainerUtil.map(items) { it.info })
       return
     }
-
     val ideDocumentHistory = IdeDocumentHistory.getInstance(project)
     for (item in items) {
-      for (it in if (isChanged) {
-        ideDocumentHistory.changePlaces
-      }
-      else {
-        ideDocumentHistory.backPlaces.filter { IdeDocumentHistory.getInstance(project).isSame(it, item.info) }
-      }) {
-        if (isChanged) {
-          ideDocumentHistory.removeChangePlace(it)
-        }
-        else {
-          ideDocumentHistory.removeBackPlace(it)
-        }
+      ContainerUtil.filter(if (isChanged) ideDocumentHistory.changePlaces else ideDocumentHistory.backPlaces) {
+        IdeDocumentHistory.getInstance(project).isSame(it, item.info)
+      }.forEach {
+        if (isChanged) ideDocumentHistory.removeChangePlace(it)
+        else ideDocumentHistory.removeBackPlace(it)
       }
     }
   }
@@ -157,9 +143,9 @@ internal class RecentLocationsDataModel(
     val text = document.getText(TextRange.create(range.startOffset, range.endOffset))
 
     val newLinesBefore = StringUtil.countNewLines(
-      StringUtil.substringBefore(text, StringUtil.trimLeading(text))!!)
+      Objects.requireNonNull<String>(StringUtil.substringBefore(text, StringUtil.trimLeading(text))))
     val newLinesAfter = StringUtil.countNewLines(
-      StringUtil.substringAfter(text, StringUtil.trimTrailing(text))!!)
+      Objects.requireNonNull<String>(StringUtil.substringAfter(text, StringUtil.trimTrailing(text))))
 
     val firstLine = document.getLineNumber(range.startOffset)
     val firstLineAdjusted = firstLine + newLinesBefore
@@ -195,12 +181,10 @@ internal class RecentLocationsDataModel(
     val startOffset = document.getLineStartOffset(startLine)
     val endOffset = document.getLineEndOffset(endLine)
 
-    return if (startOffset <= endOffset) {
+    return if (startOffset <= endOffset)
       TextRange.create(startOffset, endOffset)
-    }
-    else {
+    else
       DocumentUtil.getLineTextRange(document, line)
-    }
   }
 }
 
@@ -210,15 +194,12 @@ internal data class RecentLocationItem(
   @JvmField val linesShift: Int,
   @JvmField val ranges: Array<TextRange>
 ) {
-  override fun equals(other: Any?): Boolean {
-    if (other !is RecentLocationItem) {
-      return false
-    }
-
-    return info.file == other.info.file &&
-           linesShift == other.linesShift &&
-           text.length == other.text.length &&
-           ranges.size == other.ranges.size
+  override fun equals(other: Any?): Boolean = if (other !is RecentLocationItem) false
+  else {
+    info.file == other.info.file &&
+    linesShift == other.linesShift &&
+    text.length == other.text.length &&
+    ranges.size == other.ranges.size
   }
 
   override fun hashCode(): Int {
